ck_epoch_poll: make it safe to call ck_epoch_poll in a protected section.
This work is from jtl@FreeBSD.org. FreeBSD expects to call ck_epoch_poll
from a record that is in an active section. Previously, it was
considered an API violation to call write-side functions while in a read
section.

This is now permitted for poll as we we serialize behind the global
epoch counter.

Note that these functions are not reentrant. In the case of the
FreeBSD kernel, all these functions are called with preemption disabled.