ck_queue: remove load fences on iterators.
These were always no-op on TSO. For non-TSO, we assume
data dependent ordering. Next to data does not require ordering
but data to next does. However, a re-ordering here is benign
in that a stale load will observe an initialized NULL value.
We would rather not pay the load cost for the common case.

Write-side fences are unchanged.

From discussions with markj@freebsd.org.