ck_pr/aarch64: Specify output operands for ck_pr_md_store_*
As in commit 2f9acab, we want to specify output operand widths so that
MSAN compiler instrumentation correctly updates the shadow map.  In
particular, LLVM's implementation depends on having type information for
output operands, even when that's not otherwise necessary.  Without it,
KMSAN in FreeBSD generates false positives on aarch64.