Correct array parameter notation in CK_PR_LOAD_2()
This type of array size notation in C serves only a documentation
purpose.  That the array-size documentation provided in this
macro-templated inline function declaration is not correct for any of
the instantiations would be a minor nit except that gcc versions that
support -Wstringop-overread and/or -Wstringop-overflow use the array
sizes as guidance for when to generate those warnings (errors under
-Werror).

An alternative would be to change the parameter types to explicitly be
pointers - this fix preserves the documentation intent.