Provide a way to suppress the doc build when doing a make / make all
A variable EXTRA_BUILD is defined, with a default value of 'doc', that
establishes additional dependencies of the 'all' make target.  This
allows disabling the doc build during make / make all by passing
'EXTRA_BUILD=' on the configure command line.

This relies on the fix for name=value command line arg processing.