Generate headers in the build directory, not the source directory
The existing behavior was to generate two headers in the source
directory and copy one of them to the build directory.  Now, both
headers are generated in the build directory.  This allows a single
source tree to properly support multiple, independently configured out
of tree builds.

Note that the terminology in this configure script is currently such
that "BUILD_DIR" is actually the source dir (root of the source tree),
and "P_PWD" is the build dir (root of the build products tree).